api: procmgr: Implement minimal proc node mgr
Implements a minimal processor node registry and a simple engine
to poll for and evaluate queued measurement samples against processor
nodes found in the node registry.

If one or more matches are found between a queued measurement and a
processor node or node chain in the registry, any callbacks associated
with the processor node chain will be fired.

Signed-off-by: Kevin Townsend <kevin.townsend@linaro.org>