api: measurement: Move timestamp head of payload
Relocates the optional timestamp to the start of the payload.

This change was made to take into account situations where a
measurement buffer is allocated from the sample pool with a larger
than necessary packet size at first, to leave room for future
transformations. Maintaining the gap between the source measurement
payload and the timestamp imposes an artificial and cumbersome
constraint that is removed by placing the timestamp as the first part
of the payload, with any remaining space left as zeroes.

Signed-off-by: Kevin Townsend <kevin.townsend@linaro.org>