api: proc_mgr: Refactor node callback handlers
Updates the signature of the node callback handlers to enable
retrieval of the calling node and node instance number in the
case of node chains.

Updates the shell example to use the new callbacks, and places
the node handlers and processor node logic in a dedicated
file for easier maintenance later.

Signed-off-by: Kevin Townsend <kevin.townsend@linaro.org>